+++ src/lib/Scenes/Favorites/Components/Artists/Relay/ArtistsRenderer.tsx
@@ -15,7 +15,7 @@ export default ({ render }) => {
           }
         }
       `}
-      variables={null}
+      variables={{ count: 10 }}
       render={render}
     />
   )